ecall-keccak: move Keccak to its own module
